+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.1 - 2018-09-13
 ### Added
 - Implemented client authentication
+- Implemented toggles for primaries and indices 
 - Added status code and error checking to client requests
 
 ## 0.1.0 - 2018-08-28